A number of TM-related updates -Updated the size of BAG_TMHM_COUNT. -Ditched unnecessary TMHM_COUNT constant. -Updated the internal order of the items. -Updated the capacity of items per pocket. -Updated a handful of loop variables to u16 to account for that. -Reverted change to MAX_POCKET_ITEMS.

+++ src/rom_header_gf.c
@@ -82,7 +82,7 @@ struct GFRomHeader
     u8 bagCountItems;
     u8 bagCountKeyItems;
     u8 bagCountPokeballs;
-    u8 bagCountTMHMs;
+    u16 bagCountTMHMs;
     u8 bagCountBerries;
     u8 pcItemsCount;
     u32 pcItemsOffset;
